requirements.pipeline.txt
urllib3
requests
pathway
river
openai
groq
google-generativeai
numpy
pandas
faiss-cpu
pypdf
python-dotenv
